requirements.txt
opencv-python
numpy
python-rtmidi